refactor: move integration test into the api v1 module
