+++ eox_core/urls.py
@@ -7,8 +7,8 @@ from eox_core.api_schema import docs_ui_view
 
 app_name = 'eox_core'  # pylint: disable=invalid-name
 
-urlpatterns = [  # pylint: disable=invalid-name
-    re_path(r'^eox-info$', views.info_view),
+urlpatterns = [
+    re_path(r'^eox-info$', views.info_view, name='eox-info'),
     re_path(r'^api/', include('eox_core.api.urls', namespace='eox-api')),
     re_path(r'^data-api/', include('eox_core.api.data.v1.urls', namespace='eox-data-api')),
     re_path(r'^api-docs/$', docs_ui_view, name='apidocs-ui'),
